Add pre-commit hook to check for copyright year
This replaces the previous arclint linter for checking the copyright
year in license headers with a pre-commit hook.

This pre-commit hook is a bit different than the usual pre-commit, as we
don't want it to complain files not modified in the current year, when
run with `pre-commit run --all-files` (as we do on Gitea Actions). To
achieve that, this hook checks against the year of the last commit to a
file, if that file isn't being staged.