Add python-is-python3 to Debian CI
For the sake of vendored cxxtest. Distributions are advised to use
system cxxtest anyways.

From the package description:
This is a convenience package which ships a symlink to point the
/usr/bin/python interpreter at the current default python3. It may
improve compatibility with other modern systems, whilst breaking some
obsolete or 3rd-party software.

As this is a Debian Buster based CI and the package is only available
from Debian Bullseye forward do the symlinking manually instead.

Signed-off-by: Ralph Sennhauser <ralph.sennhauser@gmail.com>