Avoid calling random for phenotype to avoid OOS
Using the same random generator as for the simulation for previewing
leads to out-of-sync of the random generator with other clients.

This is a quick fix, in the long run one might want to provide an other
random generator for non simulation tasks.

The patch by elexis was attached by Vladislav to issue #6888

Patch-by: elexis
Closes: #6888
Signed-off-by: Ralph Sennhauser <ralph.sennhauser@gmail.com>