Refactor generate_debug_translation.py
This refactors the logic to generate debug translations to fix some
bugs, enable new functionality and be easier to read.

The bugs fixed by that are:

- for some strings the longest translation wasn't picked before
- filtering the languages to consider for the long translations didn't
  work

It also adds a new feature: In addition to the existing "long"
translation, which picks the longest translated string across all
languages, there is now a new "long2" translation, which picks the
translated string which contains the longest word across all
translations. This is helpful to check the cases where a translated
string isn't the longest one across all languages, but still causes UI
fitting problems due to the lengths of the individual words it contains.