Dynamic height for red buttons with long captions
Sizing `ModernRedButton` objects so their caption properly fits is
difficult, as contrary to `StoneButton` objects it's not possible to set
their height so that they can fit two lines of text without looking odd
when only a single line of text is necessary. In addition when
rendering a column with multiple `ModernRedButton` objects, we need all
of those buttons to have the same height, otherwise it looks odd too.

This commit tries to fix that by changing the implementation of
`distributeButtonsHorizontally()` to work with any number of buttons
and to increaste the height of buttons if any one of them has text not
fitting into a single line.

It then applies `distributeButtonsHorizontally()` to all
`ModernButtonRed` objects which need resizing to fit their captions in
different translations and ensures those buttons still fit their
surrounding layout even when their height increases.

The results is a layout where a column of buttons will have equally
distributed width. Their height will be 28px if the captions of all
buttons fits, otherwise the height of all buttons will be 42px. While
that results in different button heights depending on the selected
translation, the heights will be as small as necessary.