Add an option to toggle the match summary screen
3ab25cbd95 made the summary screen at the end of a match optional, by
asking players whether they want to show it when quitting a match. This
had several downsides. The most pressing one was that the caption of
the new button was quite short in English, but too long to fit onto
the button in many other languages. Also it added more friction for
the leaving a match, as players suddenly had to decide between two
options to do so for every match, while probablythe vast majority of
players is interested in having a look at the match summary.

To solve these downsides, but still allow developers and players to
skip the summary page, this commit removes the choice when leaving a
match to skip the summary page again. Instead it adds a configuration
option, which let's players toggle globally whether they want to see the
summary screen or not.