Add an explicit flag for unique technologies
Allows providing a `unique` boolean in a technology template, taking priority
over the previous logic for deciding whether or not the technology gets listed
as unique in the civilization overview. Previously, the (only) criteria was
depending on whether or not the civilizations that have access to it were listed
as "any of these civs" instead of "all but these civs".

Having to list excluded civs from technologies instead of the opposite, for the
only purpose of not showing as unique, leads to unclear technology templates,
the logic is confusing for contributors and modders and may lead to
oversights (see d51b4687e9  / 5f1f888a33)

A typical use case are for technologies tied to a specific unit type only trainable
by a subset of civilizations. The technology is not unique since it is
researchable by all civilizations that have access to the unit, and it is a lot
clearer to list civilizations that have access to the unit in the template,
rather than civilizations that don't. These technologies have been modified accordingly.

As a visible effect, the following technologies no longer appear as unique:
- Han farming technologies (identical effect)
- Maur wicker baskets (paired with the actual unique technology)
- Ptol tier 2 and 3 farming technologies (available in village phase, explicitly
mentioned in a civ bonus)

Fixes #7286