Move translation updates from Jenkins to Gitea
This moves the fetching of updated translations from Transifex from the
nightly-build in Jenkins to a separate Gitea Action workflow.
This is part of effort to move as much regular, but not directly
build related tasks out of Jenkins for ease of maintenance and more
flexibility.

For the new Gitea Actions workflow to work, the following secrets need
to be set:

- SVNUSER
- SVNPASS
- TX_TOKEN