Add emergency levels
This is the first small step for improved handling emergencies. It
introduces levels for emergencies because not every emergency is the same.

E.g. the AI lost a significant amount of CCs but still has e.g. 3 remaining
that's an emergency but recoverable.

The AI only has one CC left? Still recoverable, depending on the current
health.

No CC left? Basically the AI has lost