Remove the settings slide config option
This removes the settings slide config option, so it's not possible
anymore to disable the slide in/slide out of the match settings. The
rationale for that is that there is not much difference between sliding
or not and the setting itself might have been confusing for players, as
it wasn't entirely clear what it did.