Fix the line height of multi-line text
The new font rendering introduced in 734386ce9f resulted in increased
spacing between multiple lines of text.

With the old pre-computed fonts the height of a font was equal to its
cap height, so without ascenders and descenders. That required a line
spacing larger than the height of the font to account for ascenders and
descenders.

With the new font rendering logic that's not necessary anymore, because
we internally use the font height set by the font author as height,
which already includes space for ascenders and descenders. However, for
the line spacing the height difference between ascender and descender
has been used with that new logic. While that should've been fine, as
for the fonts we use "height = ascender - descender", it still resulted
in a line spacing higher than the font height, as ascender and descender
get rounded during scaling [for historical reasons by FreeType][1].

To fix that, this commit just uses the font height as line spacing as
well. While the whole line spacing logic could also be removed as part
of it, keeping it makes it straight forward to re-add support for line
heights different from the font height in future, which might be helpful
to display more condensed text when space is a premium or better
readable text with increased line height, when plenty of space is
available.

For those wondering what happens if the height of a font is different
than the height difference between ascender and descender, according to
FreeType the [font height should be used in these cases as well][2].
It's also not clear how common fonts which such properties even are, so
this might also just be a theoretical situation.

[1]:
https://freetype.org/freetype2/docs/reference/ft2-sizing_and_scaling.html#ft_size_metrics
[2]: https://freetype.org/freetype2/docs/tutorial/step2.html