Support release SM for debug build and vice versa
Except on Windows allow using any build type variant of spidermonkey to
be used with bundled spidermonkey. This was already supported for system
spidermonkey.

On one hand this reduces the build time on the other hand it makes it
somewhat practical to use the debug build of pyrogenesis not only in
exceptional cases as it's fast enough to be somewhat playable that way.

Signed-off-by: Ralph Sennhauser <ralph.sennhauser@gmail.com>