Remove precompiled.h from source files
Instead of having them in each file inject them where needed. This
allows to limit PCH to those sources where we actually make use of it in
future.

This is done in preparation for allowing the use of include-what-you-use
and would also help in adding support for other build systems.

Ref: #8086
Signed-off-by: Ralph Sennhauser <ralph.sennhauser@gmail.com>