Add exception-based OpenAL helpers (AlCall/AlcCall)
Introduce AlError/AlcError exceptions with __FILE__ & __LINE__
reporting.
- CheckALErrors / CheckAlcErrors throw on error
- AlCall / AlcCall wrap AL/ALC calls and throw on failure
- Macro namespace scoped.

Macro is need it because the template contains a Args&&... args, so it
wasn't unable to add at the end the source location with default value.

source_location isnt supported in the current OSX CICD neither linux so
still used __FILE__ & __LINE__.