Per-turn caching of template-dependent entity states
The many and slow GetEntityState calls are a main reason why large entity
selections are so laggy. This patch improves their performance in the
following way:
Big parts of an "entity state" are solely dependent on the entity's template.
Previously, those parts were all computed for each entity, even though,
among entities with the same template they would always be exactly the same.
This patch uses a cache to solve that redundancy and make sure to compute
these pieces of data only once and reuse them directly where possible.
For reasons of simplicity, this cache is strictly cleared on every turn since
refilling it is extremely quick anyway.