Add a system for scenario and campaign narratives
This patch implements a framework enabling trigger scripts to display
"narrative pages" to the player. They can be used to tell a story or
explain the historical background of a scenario or campaign.
Individual pages can be pushed to the GUI at any point during (or before)
a game without a predefined order; the GuiInterface serialises the pages
in the order it received them. When the GUI receives a new page, it shows
it immediately and pauses the game -- even in multiplayer; players
shouldn't be at a disadvantage to others just because they close the
overlay later.
If one is being displayed already, the received page is added to the end
of the list, players use the "next " and "previous" buttons to navigate
through it. After closing it, the narrative overlay can be reopened by
"Menu" (top right corner) -> "Narrative".

Each page has to define a text and optionally a title and an image as
well. The positions and sizes of the title and text can be fully
customised inside the "content container" object (visualised in
art/narratives/templates/)
The images are required to be placed in art/narratives/ where each
narrative (per scenario or campaign) should have its own subdirectory,
like the demo does. The size/placement of the image itself cannot be
customised because image dimensions are restricted to powers of two, so in
many cases it would need to be larger than what it actually shows anyway.
Instead, it is always laid exactly over the active page's parchment
texture.
Most effects one can think of can be achieved with this system.

On small screen/window resolutions the large and medium pages cannot be
displayed in the full size (they overlap with existing UI), so they and
their layout and image are scaled down manually in NarrativeOverlay.js

This patch includes a demo map called "Narrative Demo" showcasing the
feature to map makers, providing some explanations and sample code and
images.