More versatile placement of research buttons
This patch allows techs to optionally define a property 'placeBelow'
specifying a combination of class requirements. The selection panel then
tries to place their research button below the (first) training button
whose unit matches these classes (which should be just one).
This is useful for techs only affecting a single type of unit: it e.g.
allows always placing the fishing nets research button below the fishing
boats training button.

Regarding tech pairs, as explained in a comment, their two research
buttons can now be placed in four different arrangements (with
descending preference):
1. Vertically below a single training button.
2. Horizontally below two adjacent training buttons.
3. Vertically not below any training button, in the third and fourth rows
	(this is also how they were always placed previously)
4. Horizontally adjacent in the bottom row (below no training buttons).

Whenever a research button can't be placed in their preferred location
because it is already occupied, the code simply falls back to the default
position in the bottom row. This is also true for tech pairs, which
means they can now also be shown if the other right panel buttons (which
are anchored to the top) reach into the second row -- unlike previously.

Since with this patch pairs of research buttons aren't always placed
vertically, the tech pair properties 'top' and 'bottom' are also renamed
to 'first' and 'second'.